tests/wpt/web-platform-tests/html/webappapis/scripting/processing-model-2/runtime-error-cross-origin-setTimeout.html

<!doctype html>
<html>
 <head>
  <title>window.onerror - runtime error in cross-origin setTimeout</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body>
  <div id=log></div>
  <script>
    setup({allow_uncaught_exception:true});
    var t = async_test();
    var ran = false;
    window.addEventListener('error', t.step_func(e => {
        ran = true;
        assert_equals(e.error.constructor, ReferenceError);
    }));
    var script = document.createElement('script');
    script.src = location.href.replace('://', '://www1.').replace(/\/[^\/]+$/, '/support/undefined-variable-in-setTimeout.js');
    document.body.appendChild(script);
  </script>
 </body>
</html>
