LayoutTests/platform/glib/imported/w3c/web-platform-tests/html/webappapis/scripting/processing-model-2/runtime-error-cross-origin-setTimeout-expected.txt

CONSOLE MESSAGE: ReferenceError: Can't find variable: undefined_variable

PASS window.onerror - runtime error in cross-origin setTimeout
